+++ tests/Cache/CacheTest.php
@@ -1,0 +1,52 @@
+<?php
+
+use Solital\Core\Cache\Cache;
+
+class CacheTest extends \PHPUnit\Framework\TestCase
+{
+    public function testCache()
+    {
+        $cache = new Cache(true);
+
+        $list = [
+            'name' => 'Lorem Ipsum',
+            'email' => 'solital@email.com'
+        ];
+
+        if ($cache->has('list') == true) {
+            $res = $cache->get('list');
+            $this->assertIsArray($res);
+        } else {
+            $res = $cache->set('list', $list, 20);
+            $this->assertTrue($res);
+        }
+    }
+
+    public function testMultipleCache()
+    {
+        $cache = new Cache(true);
+
+        $list = [
+            'nome' => 'Harvey Specter',
+            'email' => 'specter@pearsonhardman.com'
+        ];
+
+        $list2 = [
+            'nome' => 'Louis Litt',
+            'email' => 'liitup@pearsonhardman.com'
+        ];
+
+        $return = $cache->getMultiple(['list1', 'list2']);
+
+        if (!empty($return)) {
+            $this->assertIsArray($return);
+        } else {
+            $res = $cache->setMultiple([
+                'list1' => $list,
+                'list2' => $list2
+            ], 20);
+    
+            $this->assertTrue($res);
+        }
+    }
+}